fenced type added
