+++ PSC488/psu.hpp
@@ -1,3 +1,4 @@
+#include "fencedType.hpp"
 #include "serial_comm.hpp"
 #include <QSerialPort>
 #include <exception>
@@ -6,7 +7,6 @@
 #include <qtextbrowser.h>
 #include <qwidget.h>
 #include <string>
-#include <sstream>
 
 #ifndef PSU_H
 #define PSU_H
@@ -19,47 +19,6 @@ class CommException : public std::exception {
   private:
     std::string message;
 };
-
-enum class FencedType { VOLTAGE, CURRENT, };
-
-template <FencedType X> 
-struct FencedValue {
-
-    FencedValue() : value(0) {}
-    explicit FencedValue(double value) : value(value) {}
-
-    // get rid of those weird value copy assignments
-    template <typename T> 
-    FencedValue &operator=(const T &other) = delete;
-    FencedValue &operator=(const FencedValue &other) {
-        this->value = other.value;
-        return *this;
-    };
-
-    bool operator<(const FencedValue &other) const { return value < other.value; }
-    bool operator==(const FencedValue &other) const { return value == other.value; }
-    bool operator!=(const FencedValue &other) const { return value != other.value; }
-
-    FencedValue operator-(const FencedValue &other) const { return FencedValue(value - other.value); }
-    FencedValue operator+(const FencedValue &other) const { return FencedValue(value + other.value); }
-    FencedValue operator-(double num) const { return FencedValue(value - num); }
-    FencedValue operator+(double num) const { return FencedValue(value + num); }
-
-    // conversion operator
-    operator std::string() const { 
-      /* return std::to_string(value); */ 
-      std::ostringstream out;
-      out.precision(3);
-      out << std::fixed << value;
-      return out.str();
-    }
-
-  private:
-    double value;
-};
-
-using Voltage = FencedValue<FencedType::VOLTAGE>;
-using Current = FencedValue<FencedType::CURRENT>;
 
 class Psu : public QObject {
     Q_OBJECT;
